osc/rdma: ensure bml add_procs has been called for all local procs
This fixes a bug when ob1 was not selected as the pml but osc/rdma may be
selected for an MPI window. In some cases we may use btl/sm. If this is the
case we need to ensure btl/sm knows about all the local procs (not just the
ones in the communicator). This is required for btl/sm to correctly function
at this time.

In the future btl/sm should be made more resilient.

Fixes #8434

Signed-off-by: Nathan Hjelm <hjelmn@google.com>
(cherry picked from commit 8040d05ec75ff94f2a1d84215b5c388cf957e532)