ucx: check supported transports and devices for setting priority
Add "pml_ucx_tls" parameter to control the transports to include or
exclude (with a '^' prefix). UCX will be enabled only if one of the
included transports is present, or if a non-excluded transport is
present (in case of exclude-list with a '^' prefix).

Add "pml_ucx_devices" parameter to control the devices which make UCX
component set a high priority for itself. If none of the listed devices
is present, the priority will be set to 19 - lower than ob1 priority.

Signed-off-by: Yossi Itigin <yosefe@mellanox.com>
(cherry picked from commit 562c57ce97)