Fix typo in the reorder detection.
The result was that only processes with rank >= 2 were counted in the decision,
so if the reorder was decide by the first two processes (as in the example
provided on the issue #8854) the reorder was not correctly determined.

Thanks to @bangerth for raising the issue.

Fixes #8854.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>
(cherry picked from commit 0f348fd0cc4103c9780bab39f84db934b6e79598)