configury: skip SLURM provided PMIx detection when appropriate
Do not try to detect SLURM provided PMIx when the internal
or a specific external PMIx is requested by the user.

Thanks Alexander Grund for reporting this issue.

This is a one-off commit for the v4 series since v5 and later
are unaffected since support for PMI1 and PMI2 has been dropped.

Signed-off-by: Gilles Gouaillardet <gilles@rist.or.jp>